+++ NEWS.md
@@ -7,7 +7,7 @@ News
 ### New features
 * Improved ggplot2 gene track plotting via `gg_genetracks()` to enable easy 
 layering of several ggplot2 plots above a row of gene tracks (thanks to nickhir 
-for the suggestion)
+for the suggestion).
 * For those that only want the gene tracks for their own plots, this is now 
 easier by simply not specifying `data` (or setting it to `NULL`) when calling 
 `locus()`.
@@ -26,7 +26,7 @@ to be consistent with `LDlinkR`.
 * Fixed bug when plotting LD with absent levels in `locus_ggplot()` and 
 `locus_plotly()`.
 * Fixed plots with no gene tracks (thanks to Tom Willis).
-* Genes with missing gene symbols now display the ensembl gene ID
+* Genes with missing gene symbols now display the ensembl gene ID.
 
 # locuszoomr 0.1.3
 ###### 03/11/2023